Add NoiseMap class

+++ Assets/Scripts/TerrainSettings.cs
@@ -5,6 +5,6 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "NewTerrainSettings", menuName = "Terrain Settings", order = 1)]
 public class TerrainSettings : ScriptableObject
 {
-    public float xSize = 1000f;
-    public float zSize = 1000f;
+    public int xSize = 1000;
+    public int zSize = 1000;
 }
